Search improvements: overridable settings (#391)
A search settings module that allows frankly ridiculous amounts of
override (this is merited because there are MANY settings and they
pertain to different types of thing since it's a complex area).

Defaults are set in-p[lace in the settings module.

Any or all of the defaults nested dict can be overridden in the Django
settings. Sparse dicts (i.e. only specifying the overrides) are fine; we
use ChainMap but we've extended it to be infinitely nestable.

After that, any individual setting can be overridden by and ENV var,
which can be overridden by an IndexedField's boost setting (only for
field boost parts, of course), which can then be overridden by a DB
setting. Individual settings like this are definable with a mapping of
`__` joining levels of the nested dict's keys - for example
```py
"SEARCH_EXTENDED": {
    "boost_parts":{
        "search_phrase": 2.5
    }
}
```
becomes 
```py
SEARCH_EXTENDED__boost_parts__search_phrase=2.5
```

*NB. Disabling all related field indexing temporarily. For critical
fields we have a work around with a flattened list field instead*
Related Fields causing issues with boost value retrieval, and more
generally. Hope this is more sovleable as part of cleanup once query
generation makes more sense.

---------

Co-authored-by: Cameron Lamb <cameron.lamb@digital.trade.gov.uk>